+++ check-extra.sh
@@ -9,6 +9,7 @@ cargo tomlfmt --dryrun --path yash-arith/Cargo.toml
 cargo tomlfmt --dryrun --path yash-builtin/Cargo.toml
 cargo tomlfmt --dryrun --path yash-cli/Cargo.toml
 cargo tomlfmt --dryrun --path yash-env/Cargo.toml
+cargo tomlfmt --dryrun --path yash-env-test-helper/Cargo.toml
 cargo tomlfmt --dryrun --path yash-fnmatch/Cargo.toml
 cargo tomlfmt --dryrun --path yash-quote/Cargo.toml
 cargo tomlfmt --dryrun --path yash-semantics/Cargo.toml
@@ -23,6 +24,7 @@ cargo build --package 'yash-builtin' --all-targets
 cargo build --package 'yash-builtin' --all-targets --no-default-features
 cargo build --package 'yash-cli' --all-targets
 cargo build --package 'yash-env' --all-targets
+cargo build --package 'yash-env-test-helper' --all-targets
 cargo build --package 'yash-fnmatch' --all-targets
 cargo build --package 'yash-quote' --all-targets
 cargo build --package 'yash-semantics' --all-targets
