+++ check-msrv.sh
@@ -32,6 +32,10 @@ update_workspace_member yash-env
 cargo +nightly update -Z direct-minimal-versions
 cargo +1.77.0 test --package yash-env -- $quiet
 
+update_workspace_member yash-env-test-helper
+cargo +nightly update -Z direct-minimal-versions
+cargo +1.77.0 test --package yash-env-test-helper -- $quiet
+
 update_workspace_member yash-fnmatch
 cargo +nightly update -Z direct-minimal-versions
 cargo +1.65.0 test --package yash-fnmatch -- $quiet
