+++ yash-builtin/Cargo.toml
@@ -31,4 +31,5 @@ yash-syntax = { path = "../yash-syntax", version = "0.9.0" }
 assert_matches = "1.5.0"
 futures-executor = "0.3.28"
 futures-util = { version = "0.3.28", features = ["channel"] }
+yash-env-test-helper = { path = "../yash-env-test-helper", version = "0.1.0" }
 yash-semantics = { path = "../yash-semantics", version = "0.2.0" }
