+++ yash-builtin/src/continue.rs
@@ -104,8 +104,6 @@ pub async fn main(env: &mut Env, args: Vec<Field>) -> Result {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::assert_stderr;
-    use crate::tests::assert_stdout;
     use futures_util::FutureExt;
     use std::ops::ControlFlow::Break;
     use std::rc::Rc;
@@ -117,6 +115,8 @@ mod tests {
     use yash_env::stack::Frame;
     use yash_env::Env;
     use yash_env::VirtualSystem;
+    use yash_env_test_helper::assert_stderr;
+    use yash_env_test_helper::assert_stdout;
 
     fn result_with_divert(exit_status: ExitStatus, divert: Divert) -> Result {
         let mut result = Result::new(exit_status);
