+++ yash-builtin/src/shift.rs
@@ -142,7 +142,6 @@ pub async fn main(env: &mut Env, args: Vec<Field>) -> Result {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::assert_stderr;
     use futures_util::FutureExt;
     use std::ops::ControlFlow::Break;
     use std::rc::Rc;
@@ -152,6 +151,7 @@ mod tests {
     use yash_env::stack::Builtin;
     use yash_env::stack::Frame;
     use yash_env::VirtualSystem;
+    use yash_env_test_helper::assert_stderr;
     use yash_syntax::source::Location;
 
     #[test]
