+++ yash-semantics/Cargo.toml
@@ -26,3 +26,4 @@ yash-syntax = { path = "../yash-syntax", version = "0.9.0" }
 [dev-dependencies]
 futures-executor = "0.3.28"
 futures-util = { version = "0.3.28", features = ["channel"] }
+yash-env-test-helper = { path = "../yash-env-test-helper", version = "0.1.0" }
