+++ yash-semantics/src/command.rs
@@ -85,7 +85,6 @@ impl Command for syntax::List {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::assert_stdout;
     use crate::tests::echo_builtin;
     use crate::tests::return_builtin;
     use futures_util::FutureExt;
@@ -94,6 +93,7 @@ mod tests {
     use yash_env::system::r#virtual::VirtualSystem;
     use yash_env::system::r#virtual::SIGUSR1;
     use yash_env::trap::Action;
+    use yash_env_test_helper::assert_stdout;
     use yash_syntax::source::Location;
 
     #[test]
