+++ yash-semantics/src/expansion.rs
@@ -365,11 +365,11 @@ pub async fn expand_value(
 mod tests {
     use super::*;
     use crate::tests::echo_builtin;
-    use crate::tests::in_virtual_system;
     use crate::tests::return_builtin;
     use assert_matches::assert_matches;
     use futures_util::FutureExt;
     use yash_env::variable::Scope;
+    use yash_env_test_helper::in_virtual_system;
     use yash_syntax::source::pretty::Message;
 
     #[test]
