+++ yash-semantics/src/expansion/initial/arith.rs
@@ -300,12 +300,12 @@ pub async fn expand(text: &Text, location: &Location, env: &mut Env<'_>) -> Resu
 mod tests {
     use super::*;
     use crate::tests::echo_builtin;
-    use crate::tests::in_virtual_system;
     use crate::tests::return_builtin;
     use futures_util::FutureExt;
     use yash_env::semantics::ExitStatus;
     use yash_env::system::Errno;
     use yash_env::variable::Scope::Global;
+    use yash_env_test_helper::in_virtual_system;
 
     #[test]
     fn var_env_get_variable_success() {
