+++ yash-semantics/src/runner_legacy.rs
@@ -169,8 +169,6 @@ impl<'a, 'b> ReadEvalLoop<'a, 'b> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::assert_stderr;
-    use crate::tests::assert_stdout;
     use crate::tests::echo_builtin;
     use crate::tests::return_builtin;
     use futures_util::FutureExt;
@@ -187,6 +185,8 @@ mod tests {
     use yash_env::system::r#virtual::VirtualSystem;
     use yash_env::system::r#virtual::SIGUSR1;
     use yash_env::trap::Action;
+    use yash_env_test_helper::assert_stderr;
+    use yash_env_test_helper::assert_stdout;
     use yash_syntax::source::Location;
     use yash_syntax::source::Source;
